Made a different Route for the authenticated user and started home page
