+++ src/assets/custom.css
@@ -383,3 +383,11 @@
   height: 100%;
   display: block;
 }
+
+.addsubject{
+  margin-top: 120px;
+}
+
+.chart{
+  margin-top: 120px;
+}
